# docs/swagger.yaml
basePath: /api
definitions:
...
  structs.DeletePasienResponse:
    properties:
      pesan:
        example: Pasien berhasil dihapus
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.ErrorResponse:
    properties:
      pesan:
        example: Terjadi kesalahan
        type: string
      status:
        example: error
        type: string
    type: object
...
  structs.GetPasienByIDResponse:
    properties:
      data:
        $ref: '#/definitions/structs.Pasien'
      pesan:
        example: Get pasien by ID berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
  structs.GetPasienResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/structs.Pasien'
        type: array
      pesan:
        example: Get pasien berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.Pasien:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      jenis_kelamin:
        type: string
      modified_at:
        type: string
      modified_by:
        type: string
      nama:
        type: string
      rm:
        type: string
    type: object
...
  structs.TambahPasienRequest:
    properties:
      alamat:
        type: string
      jenis_kelamin:
        type: string
      nama:
        type: string
      rm:
        type: string
    required:
    - jenis_kelamin
    - nama
    - rm
    type: object
  structs.TambahPasienResponse:
    properties:
      data:
        $ref: '#/definitions/structs.Pasien'
      pesan:
        example: Tambah pasien berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.UpdatePasienRequest:
    properties:
      alamat:
        type: string
      jenis_kelamin:
        type: string
      nama:
        type: string
      rm:
        type: string
    type: object
  structs.UpdatePasienResponse:
    properties:
      data:
        $ref: '#/definitions/structs.Pasien'
      pesan:
        example: Update pasien berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: '-'
    name: Damar
  description: 'Dokumentasi REST API Alur Obat Farmasi menggunakan Golang (Framework
    : Gin dan Database : PostgreSQL).'
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Final Project API
  version: "1.0"
paths:
...
  /api/admin-farmasi/pasien:
    get:
      description: Mengambil Semua data pasien
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil Get data pasien
          schema:
            $ref: '#/definitions/structs.GetPasienResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Semua Pasien
      tags:
      - 5. Pasien
    post:
      consumes:
      - application/json
      description: Menambahkan data pasien baru
      parameters:
      - description: Data Pasien
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.TambahPasienRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Pasien berhasil ditambahkan
          schema:
            $ref: '#/definitions/structs.TambahPasienResponse'
        "400":
          description: Gagal tambah pasien
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "409":
          description: No RM Pasien sudah digunakan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah Pasien
      tags:
      - 5. Pasien
  /api/admin-farmasi/pasien/{id}:
    delete:
      description: Menghapus data pasien
      parameters:
      - description: ID Pasien
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Pasien berhasil dihapus
          schema:
            $ref: '#/definitions/structs.DeletePasienResponse'
        "400":
          description: ID tidak valid
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Pasien tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Pasien
      tags:
      - 5. Pasien
    get:
      description: Mengambil detail pasien berdasarkan ID
      parameters:
      - description: ID Pasien
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil detail pasien
          schema:
            $ref: '#/definitions/structs.GetPasienByIDResponse'
        "400":
          description: ID Pasien tidak valid
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Pasien tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Pasien By ID
      tags:
      - 5. Pasien
    put:
      consumes:
      - application/json
      description: Mengubah data pasien
      parameters:
      - description: ID Pasien
        in: path
        name: id
        required: true
        type: integer
      - description: Data Pasien
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.UpdatePasienRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update Pasien berhasil
          schema:
            $ref: '#/definitions/structs.UpdatePasienResponse'
        "400":
          description: Gagal update pasien
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Pasien tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Pasien
      tags:
      - 5. Pasien
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...